entity_extraction/requirements.txt
google-cloud-language==1.3.0
protobuf==3.11.3
